django/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
dataclasses==0.6
Django==2.0.7
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
djongo==1.2.29
gunicorn==19.9.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
pymongo==3.7.1
pytest==3.6.3
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
